refactor: router index.js에서 분리
